add teams module
